Introduced table.h/c
Created a wrapper to store all elements of the game (Deck, Crupier and
Players). The new object will not return copies of those internal objects,
rather pointers to themselves, to enable making changes to the game trough the
Table object. The Table object is to be passed on to all
strategy/decision_making functions, so that they have access to the rest of the
game.